fyp-backend/requirements.txt
cycler==0.10.0
Django==1.11.17
django-cors-headers==3.2.1
djangorestframework==3.11.0
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.17.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
